pyproject.toml
[tool.poetry]
name = "tgbot-02-teleton"
version = "0.1.0"
description = ""
authors = ["dan.toropov <dan.toropov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.11.9"
telethon = "^1.41.2"
python-dotenv = "^1.1.1"
taskiq = "^0.11.18"
taskiq-redis = "^1.1.0"
openai = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
